change all converter volume from double to decimal

+++ src/Service.Wallet.Api/Controllers/Contracts/ExecuteSwapQuoteRequest.cs
@@ -10,8 +10,8 @@ namespace Service.Wallet.Api.Controllers.Contracts
 
         public string FromAsset { get; set; }
         public string ToAsset { get; set; }
-        public double FromAssetVolume { get; set; }
-        public double ToAssetVolume { get; set; }
+        public decimal FromAssetVolume { get; set; }
+        public decimal ToAssetVolume { get; set; }
 
         public bool IsFromFixed { get; set; }
     }
